Added session-cells.
These are similar to web-cells except they are tied to the session rather than the web frame.
They should be a simpler mechanism for storing session information.

The primary motivation for the addition is sticky notifications, which stay around after the
page has been refreshed (see next commit).

I don't intend to use session-cells very heavily just yet. Eventually I'd like to use them to
replace conventional session data storage. After that, there could be room for a generic cell
interface that encompasses web-cells, session-cells and some sort of request-cell (a temporary
variable that lasts the duration of a request).