Notifications now work correctly with AJAX.
I had to remove notification-mixin and replace it with a separate component,
notification-pane%. This makes notifications play more nicely with html-page%
and its subclasses - you can add/remove a notification without rerendering the
whole page.

I also removed the notification-related messages of notification-pane%. There's
not a lot of call for them now there are procedures to manipulate the notification
queue and they're not readily accessible from editor% or html-page% objects any more.