1. Refactored some of scaffold-internal into scaffold-common, which now contains the constant symbols representing CRUDL operations and scaffolding procedures. 2. Added the notion of a "binder" for mapping attributes onto editors.  * binder is a struct with 4 fields:   i. attributes - a list of attributes;   ii. editors - a list of editors;   iii. initialise! - a procedure that takes a struct and assigns the correct values to the editors;   iv. values - a procedure that returns a list of attribute-value pairs from the current state of the editors. Thus a binder provides procedures for converting from struct attributes into editors and back.
Binders should allow for the current scaffolder to be extended to allow:
a. multiple editors mapped to a single attribute (e.g. separate date and time fields mapped to a timestamp in the database).
b. multiple attributes mapped to a single editor (e.g. a datepicker mapping to separate day, month , and year columns in the database);
c. multiple attributes mapped to multiple editors.

Extension will be available on a per-attribute or per-type basis. Default bindings exist for the default SQL types and their default editors, which are currently:

|| Datatype       || Editor             ||
|| snooze-struct  || combo-box          ||
|| boolean        || check-box          ||
|| integer        || integer-text-field ||
|| time-utc/tai   || date-field         ||
|| strings (<128) || text-field         ||
|| strings        || text-area          ||

The extension mechanism will involve:
i. overriding the default {{{(get-attributes)}}} method, specifying the subset you want handled automatically and any you wish to handle manually (including lists of attributes);
ii. overriding the {{{(make-binding)}}} method, specifying the attributes and editors involved, a procedure from struct->editor-value and another from editor-values to a list of attribute-value pairs;

Other extensions, such as changes to the rendering mechanism, are also possible. I'll blog about them once the week is up.

"Binders" is not a great name since it suggests similarity to data-binding, which is only partially true here, as data is only transferred at request time (not live on-the-fly). Suggestions of better names are welcome!