First attempt at serializing web frames. Two hitches:
First, the web frame contains too much information because there's no equivalent of push-frame when a request begins. Even if we don't bring back parent/child relationships between frames, we should create an init-time frame to use at application setup time and a request-time frame to overlay when an app is live.

Second, because all this extra information is included in the web-frame, we hit a problem that Smoke components aren't (currently) serializable. We need a couple of wrapper syntaxes for define-serializable-class.